Update README.md
This code loads the Iris dataset, converts it to a pandas DataFrame, splits it into training and testing sets, creates a decision tree classifier with a maximum depth of 2, fits the classifier to the training data, makes predictions on the testing data, and calculates the accuracy of the model. The output shows that the model has an accuracy of 0.89, which is quite good.

Note that this is just an example, and there are many other machine learning models that could be used with the Iris dataset.